+++ src/recomment/recomment.module.ts
@@ -6,10 +6,10 @@ import { CommentModule } from 'src/comment/comment.module';
 import { AuthModule } from 'src/auth/auth.module';
 import { Comment } from 'src/comment/entities/comment.entity';
 import { User } from 'src/user/entities/user.entity';
-import { Point } from 'src/user/entities/point.entity';
 import { Post } from 'src/post/entities/post.entity';
 import { UserModule } from 'src/user/user.module';
 import { AlarmModule } from 'src/alarm/alarm.module';
+import { Point } from 'src/point/entities/point.entity';
 
 @Module({
   imports: [
