+++ src/user/user.module.ts
@@ -4,8 +4,8 @@ import { TypeOrmModule } from '@nestjs/typeorm';
 import { User } from './entities/user.entity';
 import { UserController } from './user.controller';
 import { UserService } from './user.service';
-import { Point } from './entities/point.entity';
-import { PointLog } from './entities/point-log.entity';
+import { Point } from 'src/point/entities/point.entity';
+import { PointLog } from '../point/entities/point-log.entity';
 
 @Module({
   imports: [TypeOrmModule.forFeature([User, Point, PointLog])],
